+++ metrics/influx_bridge_test.go
@@ -10,6 +10,12 @@ import (
 	influx "github.com/influxdata/line-protocol"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/collectors"
+	"github.com/stretchr/testify/assert"
+)
+
+const (
+	extraLabelName  = "extraLabel"
+	extraLabelValue = "extraValue"
 )
 
 func newTestGatherer() prometheus.Gatherer {
@@ -95,5 +101,44 @@ func TestInfluxBridgeFilter(t *testing.T) {
 	if got, want := count, 5; got != want {
 		t.Errorf("bad count: got %d, want %d", got, want)
 		fmt.Println(output)
+	}
+}
+
+func TestInfluxBridgeExtraLabels(t *testing.T) {
+	gatherer := newTestGatherer()
+	buf := new(bytes.Buffer)
+	cfg := InfluxBridgeConfig{
+		Writer:      buf,
+		Interval:    time.Minute,
+		Gatherer:    gatherer,
+		ExtraLabels: map[string]string{extraLabelName: extraLabelValue},
+	}
+	bridge, err := NewInfluxBridge(&cfg)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := bridge.Push(); err != nil {
+		t.Fatal(err)
+	}
+	output := buf.String()
+	parser := influx.NewStreamParser(strings.NewReader(output))
+	count := 0
+	for {
+		metric, err := parser.Next()
+		if err == influx.EOF {
+			break
+		}
+		count++
+		if err != nil {
+			t.Error(err)
+		}
+		foundExtraLabel := false
+		for _, tag := range metric.TagList() {
+			if tag.Key == extraLabelName && tag.Value == extraLabelValue {
+				foundExtraLabel = true
+				break
+			}
+		}
+		assert.True(t, foundExtraLabel, "extra label not found for metric")
 	}
 }